Add usage of satisfies
