+++ src/examples/type-safe-mock-data/factory-ts/nestedUserFactory.ts
@@ -1,13 +1,5 @@
-import { faker } from '@faker-js/faker';
 import { makeFactory } from 'factory.ts';
-
-export type User = {
-  firstName: string;
-  location: {
-    homeTown: string;
-    current: string;
-  };
-};
+import type { User } from './types';
 
 export const userFactory = makeFactory<User>({
   firstName: 'John',
@@ -16,11 +8,3 @@ export const userFactory = makeFactory<User>({
     homeTown: 'Munich',
   },
 });
-
-export const userFactoryWithFaker = makeFactory<User>(() => ({
-  firstName: faker.person.firstName(),
-  location: {
-    current: faker.location.city(),
-    homeTown: faker.location.city(),
-  },
-}));
