+++ src/examples/type-safe-mock-data/factory-ts/types.ts
@@ -1,0 +1,7 @@
+export type User = {
+  firstName: string;
+  location: {
+    homeTown: string;
+    current: string;
+  };
+};
